Claper.Accounts.User (Claper v1.1.0)

Functions

confirm_changeset(user)

Confirms the account by setting confirmed_at.

email_changeset(user, attrs)

A user changeset for changing the email.

It requires the email to change otherwise an error is added.

password_changeset(user, attrs, opts \\ [])

A user changeset for changing the password.

registration_changeset(user, attrs, opts \\ [])

valid_password?(arg1, password)

Verifies the password.

If there is no user or the user doesn't have a password, we call Bcrypt.no_user_verify/0 to avoid timing attacks.

validate_current_password(changeset, password)

Validates the current password otherwise adds an error to the changeset.