Claper.Embeds (Claper v1.7.0)

The Embeds context.

Functions

change_embed(embed, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking embed changes.

Examples

iex> change_embed(embed)
%Ecto.Changeset{data: %Embed{}}

create_embed(attrs \\ %{})

Creates a embed.

Examples

iex> create_embed(%{field: value})
{:ok, %Embed{}}

iex> create_embed(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_embed(event_uuid, embed)

Deletes a embed.

Examples

iex> delete_embed("123e4567-e89b-12d3-a456-426614174000", embed)
{:ok, %Embed{}}

iex> delete_embed("123e4567-e89b-12d3-a456-426614174000", embed)
{:error, %Ecto.Changeset{}}

disable_all(presentation_file_id, position)

get_embed!(id, preload \\ [])

Gets a single embed.

Raises Ecto.NoResultsError if the Embed does not exist.

Examples

iex> get_embed!(123)
%Embed{}

iex> get_embed!(456)
** (Ecto.NoResultsError)

get_embed_current_position(presentation_file_id, position)

Gets a single embed for a given position.

Examples

iex> get_embed_current_position(123, 0)
%Embed{}

list_embeds(presentation_file_id)

Returns the list of embeds for a given presentation file.

Examples

iex> list_embeds(123)
[%Embed{}, ...]

list_embeds_at_position(presentation_file_id, position)

Returns the list of embeds for a given presentation file and a given position.

Examples

iex> list_embeds_at_position(123, 0)
[%Embed{}, ...]

set_status(id, presentation_file_id, position, status)

update_embed(event_uuid, embed, attrs)

Updates a embed.

Examples

iex> update_embed("123e4567-e89b-12d3-a456-426614174000", embed, %{field: new_value})
{:ok, %Embed{}}

iex> update_embed("123e4567-e89b-12d3-a456-426614174000", embed, %{field: bad_value})
{:error, %Ecto.Changeset{}}