Claper.Events (Claper v1.4.0)

The Events context.

An activity leader is a facilitator, a user invited to manage an event.

Functions

change_activity_leader(activity_leader, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking facilitator changes.

Examples

iex> change_activity_leader(activity_leader)
%Ecto.Changeset{data: %ActivityLeader{}}

change_event(event, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking event changes.

Examples

iex> change_event(event)
%Ecto.Changeset{data: %Event{}}

count_events_month(user_id)

create_event(attrs)

Creates a event.

Examples

iex> create_event(%{field: value})
{:ok, %Event{}}

iex> create_event(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_event(event)

Deletes a event.

Examples

iex> delete_event(event)
{:ok, %Event{}}

iex> delete_event(event)
{:error, %Ecto.Changeset{}}

get_activity_leader!(id)

Gets a single facilitator.

Raises Ecto.NoResultsError if the Activity leader does not exist.

Examples

iex> get_activity_leader!(123)
%ActivityLeader{}

iex> get_activity_leader!(456)
** (Ecto.NoResultsError)

get_activity_leaders_for_event(event_id)

Gets all facilitators for a given event.

Examples

iex> get_activity_leaders_for_event!(event)
[%ActivityLeader{}, ...]

get_different_event_with_code(code, event_id)

Get a single event with the same code excluding a specific event.

Examples

iex> get_different_event_with_code("Hello", 123)
%Event{}

get_event!(id, preload \\ [])

Gets a single event.

Raises Ecto.NoResultsError if the Event does not exist.

Examples

iex> get_event!("123e4567-e89b-12d3-a456-426614174000")
%Event{}

iex> get_event!("123e4567-e89b-12d3-a456-4266141740111")
** (Ecto.NoResultsError)

get_event_with_code(code, preload \\ [])

get_event_with_code!(code, preload \\ [])

Gets a single event by code.

Raises Ecto.NoResultsError if the Event does not exist.

Examples

iex> get_event_with_code!("Hello")
%Event{}

iex> get_event_with_code!("Old event")
** (Ecto.NoResultsError)

get_managed_event!(current_user, id, preload \\ [])

Gets a single managed event.

Raises Ecto.NoResultsError if the Event does not exist.

Examples

iex> get_managed_event!(user, "123e4567-e89b-12d3-a456-426614174000")
%Event{}

iex> get_managed_event!(another_user, "123e4567-e89b-12d3-a456-426614174000")
** (Ecto.NoResultsError)

get_user_event!(user_id, id, preload \\ [])

Gets a single user's event.

Raises Ecto.NoResultsError if the Event does not exist.

Examples

iex> get_user_event!(user, "123e4567-e89b-12d3-a456-426614174000")
%Event{}

iex> get_user_event!(another_user, "123e4567-e89b-12d3-a456-426614174000")
** (Ecto.NoResultsError)

import(user_id, from_event_uuid, to_event_uuid)

Import interactions from another event

Examples

iex> import(user_id, from_event_uuid, to_event_uuid)
{:ok, %Event{}}

iex> import(user_id, from_event_uuid, to_event_uuid)
{:error, %Ecto.Changeset{}}

is_leaded_by(email, event)

Check if a user is a facilitator of a specific event.

Examples

iex> is_leaded_by("email@example.com", 123)
true

list_events(user_id, preload \\ [])

Returns the list of events of a given user.

Examples

iex> list_events(123)
[%Event{}, ...]

list_managed_events_by(email, preload \\ [])

Returns the list of events managed by a given user email.

Examples

iex> list_managed_events_by("email@example.com")
[%Event{}, ...]

update_event(event, attrs)

Updates a event.

Examples

iex> update_event(event, %{field: new_value})
{:ok, %Event{}}

iex> update_event(event, %{field: bad_value})
{:error, %Ecto.Changeset{}}