Claper.Forms (Claper v1.5.0)

The Forms context.

Functions

add_form_field(changeset)

Add an empty form field to a form changeset.

change_form(form, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking form changes.

Examples

iex> change_form(form)
%Ecto.Changeset{data: %Form{}}

change_form_submit(form_submit, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking form submit changes.

Examples

iex> change_form_submit(form_submit)
%Ecto.Changeset{data: %FormSubmit{}}

create_form(attrs \\ %{})

Creates a form.

Examples

iex> create_form(%{field: value})
{:ok, %Form{}}

iex> create_form(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_form_submit(attrs \\ %{})

Creates a form submit.

Examples

iex> create_form_submit(%{field: value})
{:ok, %FormSubmit{}}

iex> create_form_submit(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_or_update_form_submit(event_uuid, attrs)

Creates or update a FormSubmit.

Examples

iex> create_or_update_form_submit(%{field: value})
{:ok, %FormSubmit{}}

iex> create_or_update_form_submit(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_or_update_form_submit(fs, event_uuid, attrs)

delete_form(event_uuid, form)

Deletes a form.

Examples

iex> delete_form("123e4567-e89b-12d3-a456-426614174000", form)
{:ok, %Form{}}

iex> delete_form("123e4567-e89b-12d3-a456-426614174000", form)
{:error, %Ecto.Changeset{}}

delete_form_submit(event_uuid, fs)

Deletes a form submit.

Examples

iex> delete_form_submit(post, event_id)
{:ok, %FormSubmit{}}

iex> delete_form_submit(post, event_id)
{:error, %Ecto.Changeset{}}

disable_all(presentation_file_id, position)

get_form!(id, preload \\ [])

Gets a single form.

Raises Ecto.NoResultsError if the Form does not exist.

Examples

iex> get_form!(123)
%Poll{}

iex> get_form!(456)
** (Ecto.NoResultsError)

get_form_current_position(presentation_file_id, position)

Gets a single form for a given position.

Examples

iex> get_form!(123, 0)
%Form{}

get_form_submit(user_id, form_id)

Gets a single FormSubmit.

Examples

iex> get_form_submit!(321, 123)
%FormSubmit{}

get_form_submit_by_id!(id, preload \\ [])

Gets a single FormSubmit by its ID.

Raises Ecto.NoResultsError if the FormSubmit does not exist.

Examples

iex> get_form_submit_by_id!("123e4567-e89b-12d3-a456-426614174000")
%Post{}

iex> get_form_submit_by_id!("123e4567-e89b-12d3-a456-426614174123")
** (Ecto.NoResultsError)

list_form_submits(presentation_file_id)

Returns the list of form submissions for a given presentation file.

Examples

iex> list_form_submits(123)
[%FormSubmit{}, ...]

list_forms(presentation_file_id)

Returns the list of forms for a given presentation file.

Examples

iex> list_forms(123)
[%Form{}, ...]

list_forms_at_position(presentation_file_id, position)

Returns the list of forms for a given presentation file and a given position.

Examples

iex> list_forms_at_position(123, 0)
[%Form{}, ...]

remove_form_field(changeset, field)

Remove a form field from a form changeset.

set_default(id, presentation_file_id, position)

update_form(event_uuid, form, attrs)

Updates a form.

Examples

iex> update_form("123e4567-e89b-12d3-a456-426614174000", form, %{field: new_value})
{:ok, %Form{}}

iex> update_form("123e4567-e89b-12d3-a456-426614174000", form, %{field: bad_value})
{:error, %Ecto.Changeset{}}