Claper.Polls (Claper v1.4.1)

The Polls context.

Functions

add_poll_opt(changeset)

Add an empty poll opt to a poll changeset.

change_poll(poll, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking poll changes.

Examples

iex> change_poll(poll)
%Ecto.Changeset{data: %Poll{}}

create_poll(attrs \\ %{})

Creates a poll.

Examples

iex> create_poll(%{field: value})
{:ok, %Poll{}}

iex> create_poll(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_poll_vote(attrs \\ %{})

Creates a poll_vote.

Examples

iex> create_poll_vote(%{field: value})
{:ok, %PollVote{}}

iex> create_poll_vote(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_poll(event_uuid, poll)

Deletes a poll.

Examples

iex> delete_poll("123e4567-e89b-12d3-a456-426614174000", poll)
{:ok, %Poll{}}

iex> delete_poll("123e4567-e89b-12d3-a456-426614174000", poll)
{:error, %Ecto.Changeset{}}

disable_all(presentation_file_id, position)

get_poll!(id)

Gets a single poll and set percentages for each poll options.

Raises Ecto.NoResultsError if the Poll does not exist.

Examples

iex> get_poll!(123)
%Poll{}

iex> get_poll!(456)
** (Ecto.NoResultsError)

get_poll_current_position(presentation_file_id, position)

Gets a single poll for a given position.

Examples

iex> get_poll!(123, 0)
%Poll{}

get_poll_vote(user_id, poll_id)

Gets a all poll_vote.

Examples

iex> get_poll_vote!(321, 123)
[%PollVote{}]

list_polls(presentation_file_id)

Returns the list of polls for a given presentation file.

Examples

iex> list_polls(123)
[%Poll{}, ...]

list_polls_at_position(presentation_file_id, position)

Returns the list of polls for a given presentation file and a given position.

Examples

iex> list_polls_at_position(123, 0)
[%Poll{}, ...]

remove_poll_opt(changeset, poll_opt)

Remove a poll opt from a poll changeset.

set_default(id, presentation_file_id, position)

set_percentages(poll)

Calculate percentage of all poll options for a given poll.

Examples

iex> set_percentages(poll)
%Poll{}

update_poll(event_uuid, poll, attrs)

Updates a poll.

Examples

iex> update_poll("123e4567-e89b-12d3-a456-426614174000", poll, %{field: new_value})
{:ok, %Poll{}}

iex> update_poll("123e4567-e89b-12d3-a456-426614174000", poll, %{field: bad_value})
{:error, %Ecto.Changeset{}}

vote(user_id, event_uuid, poll_opts, poll_id)