ClaperWeb.Router.Helpers (Claper v1.2.1)

Module with named helpers generated from ClaperWeb.Router.

Functions

event_index_path(conn_or_endpoint, action)

event_index_path(conn_or_endpoint, action, params)

event_index_path(conn_or_endpoint, action, id, params)

event_index_url(conn_or_endpoint, action)

event_index_url(conn_or_endpoint, action, params)

event_index_url(conn_or_endpoint, action, id, params)

event_join_path(conn_or_endpoint, action)

event_join_path(conn_or_endpoint, action, params)

event_join_url(conn_or_endpoint, action)

event_join_url(conn_or_endpoint, action, params)

event_manage_path(conn_or_endpoint, action, code)

event_manage_path(conn_or_endpoint, action, code, params)

event_manage_path(conn_or_endpoint, action, code, id, params)

event_manage_url(conn_or_endpoint, action, code)

event_manage_url(conn_or_endpoint, action, code, params)

event_manage_url(conn_or_endpoint, action, code, id, params)

event_path(conn_or_endpoint, action, uuid)

event_path(conn_or_endpoint, action, uuid, params)

event_presenter_path(conn_or_endpoint, action, code)

event_presenter_path(conn_or_endpoint, action, code, params)

event_presenter_url(conn_or_endpoint, action, code)

event_presenter_url(conn_or_endpoint, action, code, params)

event_show_path(conn_or_endpoint, action, code)

event_show_path(conn_or_endpoint, action, code, params)

event_show_url(conn_or_endpoint, action, code)

event_show_url(conn_or_endpoint, action, code, params)

event_url(conn_or_endpoint, action, uuid)

event_url(conn_or_endpoint, action, uuid, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

stat_index_path(conn_or_endpoint, action, id)

stat_index_path(conn_or_endpoint, action, id, params)

stat_index_url(conn_or_endpoint, action, id)

stat_index_url(conn_or_endpoint, action, id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_confirmation_path(conn_or_endpoint, action)

user_confirmation_path(conn_or_endpoint, action, params)

user_confirmation_path(conn_or_endpoint, action, token, params)

user_confirmation_url(conn_or_endpoint, action)

user_confirmation_url(conn_or_endpoint, action, params)

user_confirmation_url(conn_or_endpoint, action, token, params)

user_registration_path(conn_or_endpoint, action)

user_registration_path(conn_or_endpoint, action, params)

user_registration_url(conn_or_endpoint, action)

user_registration_url(conn_or_endpoint, action, params)

user_reset_password_path(conn_or_endpoint, action)

user_reset_password_path(conn_or_endpoint, action, params)

user_reset_password_path(conn_or_endpoint, action, token, params)

user_reset_password_url(conn_or_endpoint, action)

user_reset_password_url(conn_or_endpoint, action, params)

user_reset_password_url(conn_or_endpoint, action, token, params)

user_session_path(conn_or_endpoint, action)

user_session_path(conn_or_endpoint, action, params)

user_session_url(conn_or_endpoint, action)

user_session_url(conn_or_endpoint, action, params)

user_settings_path(conn_or_endpoint, action, token)

user_settings_path(conn_or_endpoint, action, token, params)

user_settings_show_path(conn_or_endpoint, action)

user_settings_show_path(conn_or_endpoint, action, params)

user_settings_show_url(conn_or_endpoint, action)

user_settings_show_url(conn_or_endpoint, action, params)

user_settings_url(conn_or_endpoint, action, token)

user_settings_url(conn_or_endpoint, action, token, params)