ClaperWeb.UserAuth (Claper v1.0.0)

Functions

fetch_current_user(conn, opts)

Authenticates the user by looking into the session and remember me token.

log_in_user(conn, user, params \\ %{})

Logs the user in.

It renews the session ID and clears the whole session to avoid fixation attacks. See the renew_session function to customize this behaviour.

It also sets a :live_socket_id key in the session, so LiveView sessions are identified and automatically disconnected on log out. The line can be safely removed if you are not using LiveView.

log_out_user(conn)

Logs the user out.

It clears all session data for safety. See renew_session.

redirect_if_user_is_authenticated(conn, opts)

Used for routes that require the user to not be authenticated.

require_authenticated_user(conn, opts)

Used for routes that require the user to be authenticated.

If you want to enforce the user email is confirmed before they use the application at all, here would be a good place.